Split forecasters and persisters from main
Forecaster and Persister needed to be interfaces to allow for more
things to pull forecast data and for more ways to consume or persist
that data.  Currently Persister is fairly open, but Forecaster is still
passing around the darksky api client structs as data, this should be a
nixlight.Forecast at some point.